NXT-188 Fix failing tests
