+++ web/package.json
@@ -51,7 +51,7 @@
     "next": "^13.4.16",
     "next-logger": "^3.0.1",
     "next-plugin-node-config": "^1.0.2",
-    "next-seo": "^5.5.0",
+    "next-seo": "^5.15.0",
     "next-sitemap": "^3.1.29",
     "pino": "^6.14.0",
     "pino-mq": "^1.0.1",
