--- web/src/__mocks__/@storyblok/react.ts
@@ -1,3 +1,0 @@
-export const storyblokInit = (): null => null;
-
-export const apiPlugin = (): null => null;
